src/main/java/starting11/leftCenterBack.java
/*
 * To change this template, choose Tools | Templates
 * and open the template in the editor.
 */
package starting11;
import config.playerDetails;
/**
 *
 * @author user
 */
public class leftCenterBack implements playerDetails
{
    private String playerNumber;
    private String Name;
    private String Surname;
    private int Age;
    private int cleanSheets;
    
    private leftCenterBack() 
    {
        
    }
    
    private leftCenterBack(Builder builder)
    {
        playerNumber = builder.playerNumber;
        Name = builder.Name;
        Surname = builder.Surname;
        Age = builder.Age;
        cleanSheets = builder.cleanSheets;
    }
    
    public static class Builder
    {
        private String playerNumber;
        private String Name;
        private String Surname;
        private int Age;
        private int cleanSheets;
        
        public Builder(String playerNumber)
        {
            this.playerNumber = playerNumber;
        }
        
        public Builder Name(String value)
        {
            Name = value;
            return this;
        }
        
        public Builder Surname(String value)
        {
            Surname = value;
            return this;
        } 
        
        public Builder Age(int value)
        {
            Age = value;
            return this;
        }
        
        public Builder cleanSheets(int value)
        {
            cleanSheets = value;
            return this;
        }
            
        public leftCenterBack build()
        {
            return new leftCenterBack(this);
        }        
    }

    public String getPlayerNumber() {
        return playerNumber;
    }

    public String getName() {
        return Name;
    }

    public String getSurname() {
        return Surname;
    }

    public int getAge() {
        return Age;
    }

    public int getCleanSheets() {
        return cleanSheets;
    }
    
    public String playerSpecs(String Name, String Surname, int Age)
    {
        return("Name: "+Name+"\nSurname: "+Surname+"\nAge: "+Age);
    }
    
    @Override
    public boolean equals(Object o) {
        if (this == o) return true;
        if (o == null || getClass() != o.getClass()) return false;

        leftCenterBack lcb = (leftCenterBack) o;

        if (!playerNumber.equals(lcb.playerNumber)) return false;

        return true;
    }

    @Override
    public int hashCode() {
        return playerNumber.hashCode();
    }
}
